pyproject.toml

[project]
name = "rna_backbone_design"
version = "0.0.1"
description = "RNA-FrameFlow: Flow Matching for de novo 3D RNA Backbone Design"
authors = [
    { name = "Rishabh Anand", email = "rishabh.anand@yale.edu"},
]
readme = "README.md"
requires-python = ">= 3.11,<3.12"
license = { file = "LICENSE" }
keywords = [
    "artificial intelligence",
    "deep learning",
    "RNA design",
]

classifiers=[
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "torch>=2.4.1",
    "torchvision>=0.19.1",
    "torchaudio>=2.4.1",
    "lightning>=2.5.1",
    "hydra-core>=1.3.2",
    "torch-geometric>=2.6.1",
    "torch-scatter>=2.1.2",
    "torch-cluster>=1.6.3",
    "mdanalysis>=2.9.0",
    "mdanalysistests>=2.9.0",
    "biopandas>=0.5.1",
    "biopython>=1.85",
    "rdkit>=2024.9.6",
    "mdtraj>=1.10.3",
    "graphein>=1.7.7",
    "hydra-colorlog>=1.2.0",
    "rootutils>=1.0.7",
    "rich>=14.0.0",
    "matplotlib>=3.10.1",
    "networkx>=3.4.2",
    "gputil>=1.4.0",
    "omegaconf>=2.3.0",
    "beartype>=0.20.2",
    "jaxtyping>=0.3.1",
    "dm-tree>=0.1.8",
    "tmtools>=0.2.0",
    "pot>=0.9.5",
    "iminuit>=2.31.1",
    "tmscoring>=0.4.post0",
    "biotite>=1.2.0",
    "einops>=0.8.1",
    "ml-collections>=1.0.0",
    "wandb>=0.19.0",
    "hatchling>=1.27.0",
    "editables>=0.5",
    "setuptools>=78.1.0",
    # "flash_ipa>=0.1.0",
    # "flash-attn>=2.8.0.post2",
]

[project.urls]
Homepage = "https://pypi.org/project/rna-backbone-design/"
Repository = "https://github.com/rish-16/rna-backbone-design"

[project.optional-dependencies]
examples = []
test = [
    "pytest",
    "pytest-shard",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

# Assuming you're developing for Python 3.11
target-version = "py311"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["rna_backbone_design"]

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]

[tool.uv]
# Removed x86-only find-links for torch-geometric as GH200 is ARM64
no-build-package = ["torchvision", "torchaudio"]

[tool.uv.extra-build-dependencies]
# `flash-attn` imports `torch` during wheel build; ensure build isolation has it.
#flash-attn = ["torch"]
torch-cluster = ["torch"]
torch-scatter = ["torch"]
torch-geometric = ["torch"]

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }
torchaudio = { index = "pytorch-cu126" }
flash_ipa = { git = "https://git@github.com/flagshippioneering/flash_ipa" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
